requirements.txt
streamlit
folium
streamlit-folium
Pillow
requests
openai
python-dotenv
geocoder
pydantic
edge-tts
plotly-express